requirements.txt
# Core ML Libraries
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
scikit-learn==1.3.2
Pillow==10.1.0

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Experiment Tracking
mlflow==2.8.1

# Data Versioning
dvc==3.30.1
dvc-s3==3.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
pandas==2.1.3

# Logging & Monitoring
prometheus-client==0.19.0
